+++ lib/index.ts
@@ -100,6 +100,7 @@ class VnodeComponent implements VnodeComponent {
 
 interface Valyrian {
   (tagOrComponent: string | ValyrianComponent, props?: Props | null, children?: VnodeOrUnknown): Vnode | VnodeComponent;
+  fragment: (props: Props, children: VnodeOrUnknown[]) => VnodeOrUnknown[];
   isMounted: boolean;
   isNode: boolean;
   reservedWords: string[];
@@ -158,6 +159,10 @@ function valyrian(): Valyrian {
     } else {
       return new VnodeComponent(tagOrComponent, props || {}, children);
     }
+  };
+
+  v.fragment = (props: Props, vnodes: VnodeOrUnknown[]) => {
+    return vnodes;
   };
 
   v.isMounted = false;